requirements.txt
albumentations==1.3.0
attrs==22.1.0
gitdb==4.0.9
GitPython==3.1.27
jsonschema==4.17.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.3
matplotlib==3.6.0
matplotlib-inline==0.1.6
natsort==8.2.0
numpy==1.23.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.0.9
Pillow==9.2.0
pydub==0.25.1
python-dateutil==2.8.2
pytorch-lightning==1.6.1
scikit-image==0.19.3
scikit-learn==1.1.2
sklearn==0.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
torch==2.0.0
torch-stft==0.1.4
torchmetrics==0.9.3
torchvision==0.15.1
tqdm==4.62.3
typing_extensions==4.3.0
